airflow/dags/requirements.txt
apache-airflow
pandas
psycopg2-binary
boto3
sklearn
matplotlib